+++ lib/src/plural_resolver.dart
@@ -6,10 +6,10 @@ class PluralResolver {
   PluralResolver() : super();
 
   /// Returns the plural suffix based on [count] and presented [options].
-  String pluralize(int count, I18NextOptions options) {
+  String pluralize(Locale locale, int count, I18NextOptions options) {
     String result = '';
     if (count != 1) {
-      final number = _numberForLocale(count.abs(), options.locale);
+      final number = _numberForLocale(count.abs(), locale);
       if (number >= 0)
         result = '${options.pluralSuffix}${options.pluralSeparator}$number';
       else