+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 * Adds plural separator in I18NextOptions
 * Adds key separator in I18NextOptions
 * Adds and replaces LocalizationDataSource for ResourceStore
-* Makes `I18Next.t`'s parameters supersede the options parameter 
+* Makes `I18Next.t`'s parameters supersede the options parameter
+* Removes `Map` extension from `I18NextOptions`
+* Makes `I18NextOptions` `Diagnosticable`
 
 ## [0.0.1+4]